requirements.txt
faster-whisper>=1.1.0
anthropic>=0.42.0
tqdm>=4.66.0
yt-dlp>=2024.0.0
gradio>=5.0.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
